airflow/requirements.txt
spotipy
pandas
sqlalchemy
pg8000
kaggle
python-dotenv
pendulum
yt_dlp
minio